Fixed volume dependancy issue
the user gives a file path and the server parses it in order to get its
content and retries it from any volume and any type of document based on
the converter and returns the content to the client that requests the
references from the server.
the client's side is safe and secured.