Added support to load xml from link and fixed an old bug about encoding
Found an old bug. when the content is loaded using the content servlet
calling the server and decoding a byte code some functionalities
epecially regex matches do not work.
I fixed the bug now because even old ref documents were not leading if
we call the ui from an interface