adding some comments
I prepared al the required methods to make the merge of references based
on their positions in the document.
at first each references are parsed alone to build the array of
references then I compute their positions within the original text,
after that I merge them in an array ordered by position and finally I
parse the array and place them in the document accordingly.